requirements.txt
bottle
numpy
Pillow